Added SetTransform type functions to the renderer and TANG API. I'm going to have to change the uniform buffers / descriptor sets to one per asset, instead of the way I'm doing it now by having one single UBO / descriptor set and updating and binding it for every asset.
Changed the CMakeLists.txt script to build with W4, and fixed all new warnings. I had to disable some of them for third-party libraries such as glm and stb_image, as well as a bunch on the renderer code and logger.h

Multiple objects can now easily be loaded and have their transforms modified at runtime. The transforms are shared because I only have one set of UBO / descriptor sets