Implemented the descriptor set layout cache.
Reworked the SetLayout class to remove everything except the VkDescriptorSetLayout object so we can guarantee that an array of SetLayouts is the same as an array of VkDescriptorSetLayouts. Also removed the AddBinding function from this class because the SetLayoutCache is already in charge of allocating and tracking the descriptor set layouts. I commented out the references to AddBinding to get the project to compile cleanly and I'll take care of fixing the application properly tomorrow. Essentially, the plan is to create a DescriptorSetLayoutBuilder class that implements this AddBinding() function, which ultimately just builds up the VkDescriptorSetLayoutCreateInfo struct that we must pass into the LayoutCache to create a set layout. Sounds like a tomorrow problem....