Started splitting off the window code into it's own separate module. Turned out to be harder than I expected, since Vulkan and GLFW and more tightly bound than I thought. For example when creating the VkSurface we must call a Vulkan function and pass in a GLFWwindow* for whatever reason. This means that the renderer needs to know about GLFW unfortunately....at least that's what I think. There might be a clever workaround but I can always come back to this later.
Rework is almost done. The last part I gotta figure out is how I want the renderer and the window to communicate. The swap chain can become invalidated or outdated for whatever reason, so we must recreate it. When we do that, however, we also need to know the most up-to-date window dimensions