Implemented shader metadata generation. Also implemented shader hashing to prevent rebuilds for shaders that haven't been changed since last build. It's still missing one crucial step with include files, which happens as follows. If a glsl (include file) is edited, the build process won't recognize that every shader that includes this newly-edited glsl file must also be indirectly rebuilt because of the include.
Separately, worked on reading in the generated shader metadata in the library code to utilize it when building the set layouts for a particular pass. This also needs more work before it's possible to programatically generate the SetLayoutSummary's, but it's getting there. In order to read in the generated checksum.json file, I added a new dependency (nlohmann_json) for reading in JSON files